requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
pytest
httpx
